Migrate to infratest :tada:
