+++ infratest/scripts/run_server.sh
@@ -1,5 +1,3 @@
 #!/bin/bash
 
-# cargo build
-
 ./target/debug/packhub --generate-keys
